requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==1.0.0
pandas==2.0.3
google-cloud-bigquery==3.11.4
python-multipart==0.0.5
aiofiles==23.2.1